- reduction of 'short' noise in Blood.
Mainly using int16_t for deliberately sized struct fields, auto for temporary saves of global data and making several local variables full int.
This changes nothing substantial.

+++ source/games/blood/src/ai.cpp
@@ -1297,7 +1297,7 @@ void RecoilDude(DBloodActor* actor)
 				else aiGenDudeNewState(actor, &genDudeRecoilL);
 			}
 
-			short rState = inRecoil(pXSprite->aiState);
+			int rState = inRecoil(pXSprite->aiState);
 			if (rState > 0)
 			{
 				if (!canWalk(actor))
